ECDR-20:  Added in a parameter to the search and broker endpoint that allows the client to specify if they want the georss results to be returned as GML or as Simple GeoRSS.
Note - The only change to the AbstractAtomTransformer was adding a new method that checks the parameter value and uses that or falls back to the default if no parameter was passed in.  I accidentally formatted the file with my formatter that adds in spaces, so it looks like there is a lot of changes in there.